Split preset/camera types into dedicated file
