+++ apps/website/src/data/presets/chickenindoor.ts
@@ -10,7 +10,7 @@ import perchl from "@/assets/presets/chickenindoor/perchl.png";
 import perchr from "@/assets/presets/chickenindoor/perchr.png";
 import right from "@/assets/presets/chickenindoor/right.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const chickenindoorPresets: Record<string, Preset> = {
   home: {
