+++ apps/website/src/data/presets/chin.ts
@@ -1,6 +1,6 @@
 import home from "@/assets/presets/chin/home.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const chinPresets: Record<string, Preset> = {
   home: {
