+++ apps/website/src/data/presets/chin3.ts
@@ -1,6 +1,6 @@
 import home from "@/assets/presets/chin3/home.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const chin3Presets: Record<string, Preset> = {
   home: {
