+++ apps/website/src/data/presets/fox.ts
@@ -38,7 +38,7 @@ import treehouse from "@/assets/presets/fox/treehouse.png";
 import treehousel from "@/assets/presets/fox/treehousel.png";
 import treehouser from "@/assets/presets/fox/treehouser.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const foxPresets: Record<string, Preset> = {
   home: {
