+++ apps/website/src/data/presets/foxcorner.ts
@@ -9,7 +9,7 @@ import table from "@/assets/presets/foxcorner/table.png";
 import training2 from "@/assets/presets/foxcorner/training2.png";
 import training from "@/assets/presets/foxcorner/training.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const foxcornerPresets: Record<string, Preset> = {
   home: {
