+++ apps/website/src/data/presets/georgie.ts
@@ -7,7 +7,7 @@ import land from "@/assets/presets/georgie/land.png";
 import landz from "@/assets/presets/georgie/landz.png";
 import shore from "@/assets/presets/georgie/shore.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const georgiePresets: Record<string, Preset> = {
   home: {
