+++ apps/website/src/data/presets/georgiewater.ts
@@ -1,6 +1,6 @@
 import home from "@/assets/presets/georgiewater/home.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const georgiewaterPresets: Record<string, Preset> = {
   home: {
