+++ apps/website/src/data/presets/pushpopcrunch.ts
@@ -1,6 +1,6 @@
 import home from "@/assets/presets/pushpopcrunch/home.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const pushpopcrunchPresets: Record<string, Preset> = {
   home: {
