+++ apps/website/src/data/presets/toast.ts
@@ -8,7 +8,7 @@ import logz from "@/assets/presets/toast/logz.png";
 import rocksandwich from "@/assets/presets/toast/rocksandwich.png";
 import water from "@/assets/presets/toast/water.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const toastPresets: Record<string, Preset> = {
   home: {
