+++ apps/website/src/data/presets/wolf.ts
@@ -48,7 +48,7 @@ import trees from "@/assets/presets/wolf/trees.png";
 import waterfall from "@/assets/presets/wolf/waterfall.png";
 import wolfcornercam from "@/assets/presets/wolf/wolfcornercam.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const wolfPresets: Record<string, Preset> = {
   home: {
