+++ apps/website/src/data/presets/wolfcorner.ts
@@ -42,7 +42,7 @@ import trees from "@/assets/presets/wolfcorner/trees.png";
 import water from "@/assets/presets/wolfcorner/water.png";
 import wolfcam from "@/assets/presets/wolfcorner/wolfcam.png";
 
-import type { Preset } from "./preset";
+import type { Preset } from "../tech/cameras.types.ts";
 
 const wolfcornerPresets: Record<string, Preset> = {
   home: {
