+++ apps/website/src/data/tech/cameras.types.ts
@@ -4,3 +4,9 @@ export interface Preset {
   description: string;
   image?: StaticImageData;
 }
+
+export interface CameraData {
+  title: string;
+  group: string;
+  presets: Record<string, Preset>;
+}
